dns: the internal id should be uint64_t, not uint16_t
Probably not detected before as most DNS flows only have a
handful of transactions, but things start to break down when
the internal transaction id goes over a uint16.

+++ src/app-layer-dns-common.h
@@ -188,8 +188,8 @@ typedef struct DNSAnswerEntry_ {
 
 /** \brief DNS Transaction, request/reply with same TX id. */
 typedef struct DNSTransaction_ {
-    uint16_t tx_num;                                /**< internal: id */
-    uint16_t tx_id;                                 /**< transaction id */
+    uint64_t tx_num;                                /**< internal: id */
+    uint16_t tx_id;                                 /**< DNS transaction id */
     uint32_t logged;                                /**< flags for loggers done logging */
     uint8_t replied;                                /**< bool indicating request is
                                                          replied to. */
